Convert PNCounter to typescript.
